requirements.txt
fastapi
numpy
matplotlib
imgaug>=0.4.0
ultralytics
pillow